Remove unimplemented method from TraceConfig. Move config implementations to config package. Add tests for NoopTraceConfig.
