+++ core_impl/src/test/java/io/opencensus/trace/config/TraceConfigImplTest.java
@@ -11,11 +11,11 @@
  * limitations under the License.
  */
 
-package io.opencensus.trace;
+package io.opencensus.trace.config;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import io.opencensus.trace.config.TraceParams;
+import io.opencensus.trace.Samplers;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
